Updated to Kokkos latest version (v4.4.01) & removed deprecation warnings

+++ src/grid/iteration/kokkos-iteration.hpp
@@ -63,7 +63,7 @@ namespace schnek {
 
         template<typename... Indices>
         SCHNEK_INLINE void operator()(Indices... ind) const {
-          func(typename RangeType::LimitType{ind...});
+          func(typename RangeType::LimitType{static_cast<int>(ind)...});
         }
     };
   }  // namespace internal
